requirements.txt
numpy
pandas
scikit-learn
finta
prompt-toolkit
pyyaml
requests
schedule